--- a/src/traversals/dfs.jl
+++ b/src/traversals/dfs.jl
@@ -1,8 +1,9 @@
 # Parts of this code were taken / derived from Graphs.jl. See LICENSE for
 # licensing details.
 
-# Depth-first visit / traversal
 
+abstract type AbstractGraphVisitor end
+abstract type AbstractGraphVisitAlgorithm end
 
 #################################################
 #
@@ -10,7 +11,7 @@
 #
 #################################################
 """
-    DepthFirst
+DepthFirst
 ## Conventions in Breadth First Search and Depth First Search
 ### VertexColorMap
 - color == 0    => unseen
@@ -24,93 +25,68 @@
 mutable struct DepthFirst <: AbstractGraphVisitAlgorithm end
 
 function depth_first_visit_impl!(
-    g::AbstractGraph,      # the graph
-    stack,                          # an (initialized) stack of vertex
-    vertexcolormap::AbstractVertexMap,    # an (initialized) color-map to indicate status of vertices
-    edgecolormap::AbstractEdgeMap,      # an (initialized) color-map to indicate status of edges
-    visitor::AbstractGraphVisitor)  # the visitor
+g::AbstractGraph,      # the graph
+stack,                          # an (initialized) stack of vertex
+vertexcolormap::AbstractVertexMap,    # an (initialized) color-map to indicate status of vertices
+edgecolormap::AbstractEdgeMap,      # an (initialized) color-map to indicate status of edges
+visitor::AbstractGraphVisitor)  # the visitor
 
 
-    while !isempty(stack)
-        u, udsts, tstate = pop!(stack)
-        found_new_vertex = false
+while !isempty(stack)
+    u, udsts, tstate = pop!(stack)
+    found_new_vertex = false
 
-        while !done(udsts, tstate) && !found_new_vertex
-            v, tstate = next(udsts, tstate)
-            u_color = get(vertexcolormap, u, 0)
-            v_color = get(vertexcolormap, v, 0)
-            v_edge = Edge(u, v)
-            e_color = get(edgecolormap, v_edge, 0)
-            examine_neighbor!(visitor, u, v, u_color, v_color, e_color) #no return here
+    while !done(udsts, tstate) && !found_new_vertex
+        v, tstate = next(udsts, tstate)
+        u_color = get(vertexcolormap, u, 0)
+        v_color = get(vertexcolormap, v, 0)
+        v_edge = Edge(u, v)
+        e_color = get(edgecolormap, v_edge, 0)
+        examine_neighbor!(visitor, u, v, u_color, v_color, e_color) #no return here
 
-            edgecolormap[v_edge] = 1
+        edgecolormap[v_edge] = 1
 
-            if v_color == 0
-                found_new_vertex = true
-                vertexcolormap[v] = vertexcolormap[u] - 1 #negative numbers
-                discover_vertex!(visitor, v) || return
-                push!(stack, (u, udsts, tstate))
+        if v_color == 0
+            found_new_vertex = true
+            vertexcolormap[v] = vertexcolormap[u] - 1 #negative numbers
+            discover_vertex!(visitor, v) || return
+            push!(stack, (u, udsts, tstate))
 
-                open_vertex!(visitor, v)
-                vdsts = out_neighbors(g, v)
-                push!(stack, (v, vdsts, start(vdsts)))
-            end
-        end
-
-        if !found_new_vertex
-            close_vertex!(visitor, u)
-            vertexcolormap[u] *= -1
+            open_vertex!(visitor, v)
+            vdsts = out_neighbors(g, v)
+            push!(stack, (v, vdsts, start(vdsts)))
         end
     end
+
+    if !found_new_vertex
+        close_vertex!(visitor, u)
+        vertexcolormap[u] *= -1
+    end
+end
 end
 
 function traverse_graph!(
-    g::AbstractGraph,
-    alg::DepthFirst,
-    s::Integer,
-    visitor::AbstractGraphVisitor;
-    vertexcolormap = Dict{eltype(g),Int}(),
-    edgecolormap = DummyEdgeMap())
+g::AbstractGraph,
+alg::DepthFirst,
+s::Integer,
+visitor::AbstractGraphVisitor;
+vertexcolormap = Dict{eltype(g),Int}(),
+edgecolormap = DummyEdgeMap())
 
-    T = eltype(g)
-    vertexcolormap[s] = -1
-    discover_vertex!(visitor, s) || return
+T = eltype(g)
+vertexcolormap[s] = -1
+discover_vertex!(visitor, s) || return
 
-    sdsts = out_neighbors(g, s)
-    sstate = start(sdsts)
-    stack = [(T(s), sdsts, sstate)]
+sdsts = out_neighbors(g, s)
+sstate = start(sdsts)
+stack = [(T(s), sdsts, sstate)]
 
-    depth_first_visit_impl!(g, stack, vertexcolormap, edgecolormap, visitor)
+depth_first_visit_impl!(g, stack, vertexcolormap, edgecolormap, visitor)
 end
 
-#################################################
-#
-#  Useful applications
-#
-#################################################
 
-# Test whether a graph is cyclic
 
-mutable struct DFSCyclicTestVisitor <: AbstractGraphVisitor
-    found_cycle::Bool
-    DFSCyclicTestVisitor() = new(false)
-end
-
-function examine_neighbor!(
-    vis::DFSCyclicTestVisitor,
-    u::Integer,
-    v::Integer,
-    ucolor::Int,
-    vcolor::Int,
-    ecolor::Int)
-
-    if vcolor < 0 && ecolor == 0
-        vis.found_cycle = true
-    end
-end
-
-discover_vertex!(vis::DFSCyclicTestVisitor, v) = !vis.found_cycle
-
+# Depth-first visit / traversal
 """
     is_cyclic(g)
 
@@ -120,22 +96,7 @@ Return `true` if graph `g` contains a cycle.
 Uses DFS.
 """
 function is_cyclic end
-@traitfn is_cyclic(g::::(!IsDirected)) = ne(g) > 0
 @traitfn function is_cyclic(g::::IsDirected)
-    cmap = zeros(Int, nv(g))
-    visitor = DFSCyclicTestVisitor()
-
-    for s in vertices(g)
-        if cmap[s] == 0
-            traverse_graph!(g, DepthFirst(), s, visitor, vertexcolormap=cmap)
-        end
-        visitor.found_cycle && return true
-    end
-    return false
-end
-
-
-@traitfn function is_cyclic2(g::::IsDirected)
     T = eltype(g)
     vcolor = zeros(UInt8, nv(g))
     for v in vertices(g)
@@ -166,40 +127,14 @@ end
 end
 
 # Topological sort using DFS
+"""
+    topological_sort_by_dfs(g)
 
-mutable struct TopologicalSortVisitor{T} <: AbstractGraphVisitor
-    vertices::Vector{T}
-end
-
-function TopologicalSortVisitor(n::T) where T<:Integer
-    vs = Vector{T}()
-    sizehint!(vs, n)
-    return TopologicalSortVisitor(vs)
-end
-
-function examine_neighbor!(visitor::TopologicalSortVisitor, u::Integer, v::Integer, ucolor::Int, vcolor::Int, ecolor::Int)
-    (vcolor < 0 && ecolor == 0) && error("The input graph contains at least one loop.")
-end
-
-function close_vertex!(visitor::TopologicalSortVisitor, v::Integer)
-    push!(visitor.vertices, v)
-end
-
-function topological_sort_by_dfs(g::AbstractGraph)
-    nvg = nv(g)
-    cmap = zeros(Int, nvg)
-    visitor = TopologicalSortVisitor(nvg)
-
-    for s in vertices(g)
-        if cmap[s] == 0
-            traverse_graph!(g, DepthFirst(), s, visitor, vertexcolormap=cmap)
-        end
-    end
-
-    reverse(visitor.vertices)
-end
-
-function topological_sort_by_dfs2(g::AbstractGraph)
+Return a [toplogical sort](https://en.wikipedia.org/wiki/Topological_sorting) of a directed
+graph `g` as a vector of vertices in topological order.
+"""
+function toplogical_sort_by_dfs end
+@traitfn function topological_sort_by_dfs(g::::IsDirected)
     T = eltype(g)
     vcolor = zeros(UInt8, nv(g))
     verts = Vector{T}()
@@ -231,47 +166,13 @@ function topological_sort_by_dfs2(g::AbstractGraph)
     return reverse(verts)
 end
 
-mutable struct TreeDFSVisitor{T} <:AbstractGraphVisitor
-    tree::DiGraph
-    predecessor::Vector{T}
-end
-
-TreeDFSVisitor(n::T) where T<:Integer = TreeDFSVisitor(DiGraph(n), zeros(T, n))
-
-function examine_neighbor!(visitor::TreeDFSVisitor, u::Integer, v::Integer, ucolor::Int, vcolor::Int, ecolor::Int)
-    if (vcolor == 0)
-        visitor.predecessor[v] = u
-    end
-    return true
-end
-
 """
     dfs_tree(g, s)
 
 Return an ordered vector of vertices representing a directed acylic graph based on
 depth-first traversal of the graph `g` starting with source vertex `s`.
 """
-function dfs_tree(g::AbstractGraph, s::Integer)
-    nvg = nv(g)
-    visitor = TreeDFSVisitor(nvg)
-    traverse_graph!(g, DepthFirst(), s, visitor)
-    # visitor = traverse_dfs(g, s, TreeDFSVisitor(nvg))
-    h = DiGraph(nvg)
-    for (v, u) in enumerate(visitor.predecessor)
-        if u != 0
-            add_edge!(h, u, v)
-        end
-    end
-    return h
-end
-
-# """
-# dfs_tree(g, s)
-
-# Return an ordered vector of vertices representing a directed acylic graph based on
-# depth-first traversal of the graph `g` starting with source vertex `s`.
-# """
-# dfs_tree(g::AbstractGraph, s::Integer; dir=:out) = tree(dfs_parents(g, s; dir=dir))
+dfs_tree(g::AbstractGraph, s::Integer; dir=:out) = tree(dfs_parents(g, s; dir=dir))
 
 """
 dfs_parents(g, s[; dir=:out])
@@ -287,35 +188,6 @@ dfs_parents(g::AbstractGraph, s::Integer; dir=:out) =
 (dir == :out) ? _dfs_parents(g, s, out_neighbors) : _dfs_parents(g, s, in_neighbors)
 
 function _dfs_parents(g::AbstractGraph, s::Integer, neighborfn::Function)
-    T = eltype(g)
-    parents = zeros(T, nv(g))
-
-    seen = falses(nv(g))
-    S = Vector{T}([s])
-    seen[s] = true
-    parents[s] = s
-    while !isempty(S)
-        v = S[end]
-        u = 0
-        for n in neighborfn(g, v)
-            if !seen[n]
-                u = n
-                break
-            end
-        end
-        if u == 0
-            pop!(S)
-        else
-            seen[u] = true
-            push!(S, u)
-            parents[u] = v
-        end
-    end
-    return parents
-end
-dfs_tree2(g::AbstractGraph, s::Integer; dir=:out) = tree(dfs_parents(g, s; dir=dir))
-
-function _dfs_parents3(g::AbstractGraph, s::Integer, neighborfn::Function)
     T = eltype(g)
     parents = zeros(T, nv(g))
 
@@ -342,4 +214,3 @@ function _dfs_parents3(g::AbstractGraph, s::Integer, neighborfn::Function)
     end
     return parents
 end
-dfs_tree3(g::AbstractGraph, s::Integer; dir=:out) = tree(_dfs_parents3(g, s, out_neighbors))