--- a/src/traversals/dfs.jl
+++ b/src/traversals/dfs.jl
@@ -90,7 +90,7 @@ depth-first traversal of the graph `g` starting with source vertex `s`.
 dfs_tree(g::AbstractGraph, s::Integer; dir=:out) = tree(dfs_parents(g, s; dir=dir))
 
 """
-dfs_parents(g, s[; dir=:out])
+    dfs_parents(g, s[; dir=:out])
 
 Perform a depth-first search of graph `g` starting from vertex `s`.
 Return a vector of parent vertices indexed by vertex. If `dir` is specified,