--- a/src/traversals/maxadjvisit.jl
+++ b/src/traversals/maxadjvisit.jl
@@ -3,7 +3,6 @@
 
 # Maximum adjacency visit / traversal
 
-
 #################################################
 #
 #  Maximum adjacency visit