--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -71,7 +71,6 @@ tests = [
     "shortestpaths/dijkstra",
     "shortestpaths/floyd-warshall",
     "shortestpaths/yen",
-    "traversals/graphvisit",
     "traversals/bfs",
     "traversals/parallel_bfs",
     "traversals/bipartition",