Benchmarks below.

To do:
- [x] `src/traversals/dfs.jl`
- [ ] `src/connectivity.jl`
- [ ] `src/biconnect/articulation.jl` (optional, since it doesn't use the dfs `traverse_graph!`)
- [ ] `src/biconnect/biconnect.jl` (optional, since it doesn't use the dfs `traverse_graph!`)