@sbromberger 2017-08-17
First benchmarks: `dfs_tree`.

```
# largest CC of a kronecker graph
julia> g = loadgraph("/tmp/testg.jlg")
{31206, 3679818} undirected simple Int64 graph

# current master
julia> @benchmark LightGraphs.dfs_tree($g, 1)
BenchmarkTools.Trial:
  memory estimate:  22.80 MiB
  allocs estimate:  302223
  --------------
  minimum time:     291.420 ms (3.38% GC)
  median time:      356.236 ms (2.96% GC)
  mean time:        358.957 ms (3.00% GC)
  maximum time:     422.605 ms (2.51% GC)
  --------------
  samples:          14
  evals/sample:     1

# with a bitvector for seen state
julia> @benchmark LightGraphs.dfs_tree2($g, 1)
BenchmarkTools.Trial:
  memory estimate:  13.72 MiB
  allocs estimate:  177343
  --------------
  minimum time:     102.124 ms (0.00% GC)
  median time:      115.666 ms (6.15% GC)
  mean time:        121.500 ms (4.95% GC)
  maximum time:     164.035 ms (5.83% GC)
  --------------
  samples:          42
  evals/sample:     1

# with a vector of Bool for seen state
julia> @benchmark LightGraphs.dfs_tree3($g, 1)
BenchmarkTools.Trial:
  memory estimate:  13.75 MiB
  allocs estimate:  177340
  --------------
  minimum time:     55.939 ms (0.00% GC)
  median time:      65.307 ms (10.29% GC)
  mean time:        64.737 ms (7.89% GC)
  maximum time:     83.763 ms (0.00% GC)
  --------------
  samples:          78
  evals/sample:     1
```