@sbromberger 2017-08-17
```
julia> d = LightGraphs.dfs_tree2(g, 1)
{1000000, 999999} directed simple Int64 graph

julia> @benchmark LightGraphs.topological_sort_by_dfs($d)
BenchmarkTools.Trial:
  memory estimate:  92.92 MiB
  allocs estimate:  2000028
  --------------
  minimum time:     242.322 ms (34.73% GC)
  median time:      327.402 ms (46.63% GC)
  mean time:        440.248 ms (61.91% GC)
  maximum time:     733.249 ms (74.57% GC)
  --------------
  samples:          12
  evals/sample:     1

julia> @benchmark LightGraphs.topological_sort_by_dfs2($d)
BenchmarkTools.Trial:
  memory estimate:  26.58 MiB
  allocs estimate:  45
  --------------
  minimum time:     76.576 ms (3.77% GC)
  median time:      79.960 ms (3.78% GC)
  mean time:        81.170 ms (3.73% GC)
  maximum time:     93.135 ms (3.21% GC)
  --------------
  samples:          62
  evals/sample:     1
```