@sbromberger 2017-08-22
I will be cleaning this up and merging tomorrow unless there are any objections.